Did maintenance on the file structure, updated platform detection logic, and changed core.h to base.h, fixed a few warnings

+++ Waffle/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -57,7 +57,7 @@ namespace Waffle {
 		for (const auto& element : layout)
 		{
 			glEnableVertexAttribArray(index);
-			glVertexAttribPointer(index, element.GetComponentCount(), ShaderDataTypeToOpenGLBaseType(element.Type), element.Normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (const void*)element.Offset);
+			glVertexAttribPointer(index, element.GetComponentCount(), ShaderDataTypeToOpenGLBaseType(element.Type), element.Normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (const void*)(intptr_t)element.Offset);
 			index++;
 		}
